% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LanZhou.R
\docType{data}
\name{LanZhou}
\alias{LanZhou}
\title{solar elevation angles of LanZhou.}
\format{
A data frame with 365 rows and 2 variables:
\describe{
  \item{Date}{date_time from "2021/01/01 12:00:00" to "2021/12/31 12:00:00"}
  \item{H}{The solar elevation angles at 12 o 'clock in Lanzhou}
  ...
}
}
\usage{
LanZhou
}
\description{
A dataset containing the annual solar elevation angles at 12 o 'clock in Lanzhou.
}
\keyword{datasets}
