% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSolarElevationAngles.R
\name{SolarAltitudeAngle}
\alias{SolarAltitudeAngle}
\title{Calculate the solar elevation angles.}
\usage{
SolarAltitudeAngle(date, lat)
}
\arguments{
\item{date}{datetime-object or String vector}

\item{lat}{Numeric vector}
}
\value{
Numeric vector of solar elevation angles
}
\description{
Calculate the solar elevation angles from the input data.
}
\details{
Calculate the solar elevation angles of the location based on the input date and dimension.
}
\examples{
SolarAltitudeAngle("2021/4/15 14:00:00",36.05)
SolarAltitudeAngle("2021/6/22 12:00:00",40)
\dontrun{
SolarAltitudeAngle("2021/6/22",40)
}
}
